+++ src/Types.ts
@@ -623,5 +623,6 @@ export type FinanceKitError =
 
 export interface SubmissionData {
     phoneNumber?: string;
+    dateOfBirth?: string;
 };